Move subgraph preview module to the target shape; test result mapping with a Tower mock
Address two review findings on the subgraph-preview implementation:

- Module shape: fold types.rs and the GraphQLQuery derives into mod.rs,
  leaving service.rs with only the tower::Service wrappers, per
  CLAUDE.md's target shape for new rover-client operations.
- The six result_* tests exercising ComposeAndFilterPreviewResult's
  pure field mapping now use a Tower mock (rover_tower::mock_service!)
  instead of spinning up a real MockServer, since they were never
  testing wire-level HTTP behavior.
